ensure consistent mod action id from Atom feed
When feedparser.parse fetches Atom from a URL, it "resolves" ids of all
feed entries (like many other URI elements) and turns them into full
URLs. This results in id values different from JSON feed and is not
what we want, so hack it to always extract the real id in the form
`ModAction_<uuid>` (length 46).

https://feedparser.readthedocs.io/en/latest/resolving-relative-links.html

Note that since 5fa9aa3 feedparser.parse is not fetching the feed
itself and istead is given a string that was already fetched by
requests.get. Therefore its odd id conversion behavior is no longer
used. But just in case we will use it again, ensure consistent ids
in the converter function.