save timestamp of newest seen mod action id
To handle various edge cases when fetched items have unexpected
timestamps compared to newest seen mod action timestamp, we will need
to know the latter. But newest seen mod action may not exist in the db
because it did not pass the filters. For example, if it was an
`editflair` action, its id was recorded as newest seen, but it was
filtered out and was not saved. To have its timestamp, save it in the
meta table along the id.