fix second run after initial run with empty feed
The bug went as follows:

- first run initializes the db and sets newest_modaction_timestamp to ""
- if the first run did not fetch any data, the timestamp remains ""
- second run tries to parse "" as int and fails

The buggy line 236 was introduced in b1e847b but the bug was highly
improbable because table row in question was always set at first run.
Then it got more probable in 6192190 when that line was extracted
for reuse by other meta values. Finally, the bug started crashing the
program cb09637 when it was used for another meta value with int type.

I should start writing unit tests some day.