rename reason field to details
"details" is more generic and captures any mod action details that
didn't fit in other fields. In fact, reason field is already used this
way. Will need to updated the modlog spec too, but this rename does not
affect spec compliance.

Bump schema version since it is a breaking change.