fix timestamp conversion from JSON feed
First run worked but subsequent runs crashed on reading the
newest_modacion_id value from the database, specifically parsing it from
str to int. It failed because a float value was written to the db as a
str, because Reddit JSON feeds return floats in created_utc fields.

The bug was introduced in ebbceb8 when I changed ModAction.date field
(with datetime values) to .timestamp (with int values). Atom modaction
converter received and explicit int() conversion but not the JSON one.

I missed the bug because I was testing agains local .json files (super
fast and convenient) that were formatted with the jq tool, which
happened to convert floats with zeroed decimal values to integers. The
bug did not manifest at that point since there was no reading from the
db.

The bug got real and started crashing the program in cb09637 when saving
and loading of the newest seen timestamp was added. I still did not hit
it with my local .json files.

Finally, occasional testing with real Reddit requests revealed the bug.