save raw JSON mod actions
Currently only a subset of Reddit response is used and saved in the
redditmodlog database. When `json_save_raw` is `true` in config.ini, raw
JSON mod action objects (unmodified and unfiltered) are stored in a
separate database file.

This feature is only active in `mode=json`.

If in the future we decide to extract and use more parts of Reddit
modlog response, it will be useful to have full raw past data available.
Note that Reddit allows to get the mod log only ~2 months into the past.

This change uses a sligthly improved approach to database handling.
First, use contextlib.closing to ensure the db is closed at the end of
the function. Second, store database schema version in SQLite PRAGMA
user_version that exists for this very purpose.

https://sqlite.org/pragma.html#pragma_user_version