use Markdown to print compact object links
Originally I was printing full object links to support Matrix chat
clients that can only show plaintext but no HTML or Markdown. But
abusing Markdown links a bit allows to have both short links in capable
clients and _some_ way to access the object in restricted clients.

If we put post and comment IDs as link text, it allows to reconstruct
the full link even if these IDs is all you see in a plaintext client.