fix get_db_value returning nothing for 0 int value
When integer 0 is coming from the db, get_db_value must return 0.
Instead it returned None because `bool(0)` is False. So if `row = (0,)`,
`row and row[0]` evaluates to False.

The bug has not manifested so far because the function was never called
with an integer value 0 from the db. This is the case for
`PRAGMA user_version` query for an empty (new) .sqlite file, where we
expect version to be 0 and not None.

This is a good general lesson to "pass bool-ish values through
unchanged" and only convert them to bools when really needed.

Because get_db_value can now return an empty string, add an int
converter that can handle that.