remove target comment previews
8f477cf introduced posting the full comment body for mod actions of type
`distinguish`. It is a useful to see extra context directly in the
modlog room, but this naive implementation did not handle certain cases:

- comments with blockquotes (>) were rendered poorly
- comments with multiple paragraphs were rendered poorly
- too long comments would spam the modlog
- user-generated input requires sanitization

The last one is a security issue. The risk of the removed implementation
was low because it only posted distinguished comments. Only subreddit
mods can distinguish comments, and only their own. So this would only be
_possibly_ exploitable by malicious mods. Still, I don't want to deai
with all of this now as it might take too much time and there are more
important things.